exercices/tp4/requirements.txt
fastapi
pydantic
langchain
langchain-google-genai
langchain-core
python-dotenv
streamlit
requests
google-cloud-storage
langchain-google-vertexai
langchain-unstructured
langchain-community
faiss-cpu  # or use faiss-gpu for GPU support
unstructured[pptx]
uvicorn